Collision Shape was Disabeled
It totally was just that the collision shape was disabled, but because we made the enemy local to the level scene, the changes we made to the enemy scene didn't propagate. Deleting the enemy and re-importing it as a reference to the scene fixes the issue.

Now you just have to figure out why it's flying lol.